+++ src/QcloudApi/Common/Request.php
@@ -23,7 +23,7 @@ class QcloudApi_Common_Request
      * $_version
      * @var string
      */
-    protected static $_version = 'SDK_PHP_2.0.5';
+    protected static $_version = 'SDK_PHP_2.0.6';
 
     /**
      * $_timeOut